+++ packages/mds-collector-backend/client/index.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { RpcClient, RpcRequestOptions, RpcRequestWithOptions } from '@mds-core/mds-rpc-common'
+import { RpcClient, RpcRequest, RpcRequestOptions } from '@mds-core/mds-rpc-common'
 import { ServiceClient } from '@mds-core/mds-service-helpers'
 import {
   CollectorMessageDomainModel,
@@ -29,10 +29,9 @@ const CollectorServiceRpcClient = RpcClient(CollectorServiceRpcDefinition, {
 })
 
 export const CollectorServiceClientFactory = (options: RpcRequestOptions = {}): ServiceClient<CollectorService> => ({
-  registerMessageSchema: (...args) =>
-    RpcRequestWithOptions(options, CollectorServiceRpcClient.registerMessageSchema, args),
-  getMessageSchema: (...args) => RpcRequestWithOptions(options, CollectorServiceRpcClient.getMessageSchema, args),
-  writeSchemaMessages: (...args) => RpcRequestWithOptions(options, CollectorServiceRpcClient.writeSchemaMessages, args)
+  registerMessageSchema: (...args) => RpcRequest(options, CollectorServiceRpcClient.registerMessageSchema, args),
+  getMessageSchema: (...args) => RpcRequest(options, CollectorServiceRpcClient.getMessageSchema, args),
+  writeSchemaMessages: (...args) => RpcRequest(options, CollectorServiceRpcClient.writeSchemaMessages, args)
 })
 
 export const CollectorServiceClient = CollectorServiceClientFactory()
@@ -44,13 +43,12 @@ export const CollectorServiceSchemaClient = (
 ) => {
   return {
     registerMessageSchema: () =>
-      RpcRequestWithOptions(options, CollectorServiceRpcClient.registerMessageSchema, [schema_id, schema]),
-    getMessageSchema: () => RpcRequestWithOptions(options, CollectorServiceRpcClient.getMessageSchema, [schema_id]),
+      RpcRequest(options, CollectorServiceRpcClient.registerMessageSchema, [schema_id, schema]),
+    getMessageSchema: () => RpcRequest(options, CollectorServiceRpcClient.getMessageSchema, [schema_id]),
     writeSchemaMessages: (
       provider_id: CollectorMessageDomainModel['provider_id'],
       messages: Array<CollectorMessageDomainModel['message']>
-    ) =>
-      RpcRequestWithOptions(options, CollectorServiceRpcClient.writeSchemaMessages, [schema_id, provider_id, messages])
+    ) => RpcRequest(options, CollectorServiceRpcClient.writeSchemaMessages, [schema_id, provider_id, messages])
   }
 }