+++ packages/mds-compliance-service/client/index.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { RpcClient, RpcRequestOptions, RpcRequestWithOptions } from '@mds-core/mds-rpc-common'
+import { RpcClient, RpcRequest, RpcRequestOptions } from '@mds-core/mds-rpc-common'
 import { ServiceClient } from '@mds-core/mds-service-helpers'
 import { ComplianceService, ComplianceServiceDefinition } from '../@types'
 
@@ -25,26 +25,23 @@ const ComplianceServiceRpcClient = RpcClient(ComplianceServiceDefinition, {
 
 // What the API layer, and any other clients, will invoke.
 export const ComplianceServiceClientFactory = (options: RpcRequestOptions = {}): ServiceClient<ComplianceService> => ({
-  getComplianceSnapshot: (...args) =>
-    RpcRequestWithOptions(options, ComplianceServiceRpcClient.getComplianceSnapshot, args),
+  getComplianceSnapshot: (...args) => RpcRequest(options, ComplianceServiceRpcClient.getComplianceSnapshot, args),
   getComplianceSnapshotsByTimeInterval: (...args) =>
-    RpcRequestWithOptions(options, ComplianceServiceRpcClient.getComplianceSnapshotsByTimeInterval, args),
+    RpcRequest(options, ComplianceServiceRpcClient.getComplianceSnapshotsByTimeInterval, args),
   getComplianceSnapshotsByIDs: (...args) =>
-    RpcRequestWithOptions(options, ComplianceServiceRpcClient.getComplianceSnapshotsByIDs, args),
-  createComplianceSnapshot: (...args) =>
-    RpcRequestWithOptions(options, ComplianceServiceRpcClient.createComplianceSnapshot, args),
+    RpcRequest(options, ComplianceServiceRpcClient.getComplianceSnapshotsByIDs, args),
+  createComplianceSnapshot: (...args) => RpcRequest(options, ComplianceServiceRpcClient.createComplianceSnapshot, args),
   createComplianceSnapshots: (...args) =>
-    RpcRequestWithOptions(options, ComplianceServiceRpcClient.createComplianceSnapshots, args),
+    RpcRequest(options, ComplianceServiceRpcClient.createComplianceSnapshots, args),
   getComplianceViolationPeriods: (...args) =>
-    RpcRequestWithOptions(options, ComplianceServiceRpcClient.getComplianceViolationPeriods, args),
-  getComplianceViolation: (...args) =>
-    RpcRequestWithOptions(options, ComplianceServiceRpcClient.getComplianceViolation, args),
+    RpcRequest(options, ComplianceServiceRpcClient.getComplianceViolationPeriods, args),
+  getComplianceViolation: (...args) => RpcRequest(options, ComplianceServiceRpcClient.getComplianceViolation, args),
   getComplianceViolationsByTimeInterval: (...args) =>
-    RpcRequestWithOptions(options, ComplianceServiceRpcClient.getComplianceViolationsByTimeInterval, args),
+    RpcRequest(options, ComplianceServiceRpcClient.getComplianceViolationsByTimeInterval, args),
   createComplianceViolation: (...args) =>
-    RpcRequestWithOptions(options, ComplianceServiceRpcClient.createComplianceViolation, args),
+    RpcRequest(options, ComplianceServiceRpcClient.createComplianceViolation, args),
   createComplianceViolations: (...args) =>
-    RpcRequestWithOptions(options, ComplianceServiceRpcClient.createComplianceViolations, args)
+    RpcRequest(options, ComplianceServiceRpcClient.createComplianceViolations, args)
 })
 
 export const ComplianceServiceClient = ComplianceServiceClientFactory()