+++ packages/mds-policy-author/handlers/update-policy.ts
@@ -27,7 +27,6 @@ export const UpdatePolicyHandler = async (
   const updatePolicy = req.body
   try {
     if (updatePolicy.publish_date) throw new ValidationError('publish_date cannot be set via policy editing endpoint')
-
     const policy = await PolicyServiceClient.editPolicy(updatePolicy)
     return res.status(200).send({ version: res.locals.version, data: { policy } })
   } catch (error) {