+++ packages/mds-stream/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @mds-core/mds-stream
 
+## 0.1.45
+
+### Patch Changes
+
+- Updated dependencies [a156f493]
+  - @mds-core/mds-types@0.7.0
+  - @mds-core/mds-utils@0.2.8
+
 ## 0.1.44
 
 ### Patch Changes